correccion de tipo

+++ WhittedRayTracing/constantes/parametros.h
@@ -7,6 +7,6 @@ const int IMAGEN_WIDTH = 1280;
 const int IMAGEN_HEIGHT = 720;
 
 const int PROFUNDIDAD_MAXIMA = 5; // Profunidad maxima en recursion de ray tracing
-const int EPSILON = 0.001; // Para mitigar colisiones de rayos con su lugar de partida
+const float EPSILON = 0.001f; // Para mitigar colisiones de rayos con su lugar de partida
 
 #endif
